Make direnv local postgres creation more manual
An issue here is that direnv often gets to run prior to asdf / asdf-direnv getting everything installed and paths set,
so running initdb and friends happens on the system version of postgres, not the local addf-postgres version with the
correct direnv paths.  There's probably a way to be slick and trigger asdf etc. here, but it's also fine, I think to
be deliberate on what is likely the only time we're creating this pg instance and database for this project.